+++ forge/CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to the format [MCVERSION-MAJORMOD.MAJORAPI.MINOR.PATCH](https://mcforge.readthedocs.io/en/1.16.x/conventions/versioning/).
 
+## [1.16.4-0.3] - 2020.12.30
+### Added
+- Added configuration options for maxBlocksBase and maxDistanceBase for granting vein mining
+abilities without the enchantment
+
 ## [1.16.4-0.2] - 2020.12.29
 ### Fixed
 - Fixed NullPointerException with Quark [#1](https://github.com/TheIllusiveC4/VeinMining/issues/1)